refactor: move modal modules to shared, move image-viewer modules & sidebar modules to features
